refactor(ci): share path classification
